Move code into library.
