+++ unpackerpoller/helpers.go
@@ -1,4 +1,4 @@
-package main
+package unpackerpoller
 
 import (
 	"io/ioutil"
