+++ unpackerpoller/pollers.go
@@ -1,4 +1,4 @@
-package main
+package unpackerpoller
 
 import (
 	"log"
